+++ qiskit_ibm_runtime/executor_estimator/pec/prepare_pec.py
@@ -36,7 +36,7 @@ from ...quantum_program import QuantumProgram
 from ...quantum_program.quantum_program import SamplexItem
 from ..trex_utils import create_trex_calibration_circuit
 from ..utils import box_circuit, compute_samplex_arguments, make_samplex_arguments
-from .pec_utils import calculate_gamma, calculate_pec_twirling_shots
+from .utils import calculate_gamma, calculate_pec_twirling_shots
 
 logger = logging.getLogger(__name__)
